pyproject.toml
[project]
name = "news-recommendation-llm"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "YadaYuki", email = "yada.yuki@fuji.waseda.jp" }
]
dependencies = ["polars~=0.18.2", "pandas~=2.0.2", "pyarrow~=12.0.0", "requests~=2.31.0", "tqdm~=4.65.0", "pydantic~=1.10.9", "notebook~=6.5.4", "matplotlib~=3.7.1", "scikit-learn~=1.2.2", "implicit~=0.7.0", "h5py~=3.9.0", "torch~=2.0.1", "transformers~=4.30.2", "accelerate~=0.21.0", "hydra-core~=1.3.2"]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.rye]
managed = true
dev-dependencies = ["ruff~=0.0.270", "black~=23.3.0", "mypy~=1.3.0", "pytest~=7.3.1", "types-requests~=2.31.0.1"]
[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 119

[tool.ruff]
line-length = 119

[tool.ruff.isort]
lines-after-imports = 2

[tool.mypy]
python_version = 3.11
disallow_untyped_defs=true
